+++ CHANGES.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- Refactor unicode normalize #315 @mosuka
+
 ## 0.23.0 (2023-02-23)
 - Bump up version to 0.23.0 #314 @mosuka
 - Add filter feature #313 @mosuka